requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
pydantic==2.6.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.26.0
langchain==0.1.13
langchain-core==0.1.33
langchain-ollama==0.1.1
langgraph==0.0.66
urllib3==2.1.0
aiosqlite==3.1.1
geopy==2.4.1
python >= 3.10